enhancement: move welcome view into app setup and remove crash reporting view
