+++ packages/shared/lib/core/router/enums/onboarding/app-setup-route.enum.ts
@@ -1,5 +1,5 @@
 export enum AppSetupRoute {
     Legal = 'legal',
-    CrashReporting = 'crashReporting',
     LanguageAndAppearance = 'languageAndAppearance',
+    Welcome = 'welcome',
 }
