+++ packages/shared/routes/onboarding/views/app-setup/AppSetupRouter.svelte
@@ -1,17 +1,17 @@
 <script lang="typescript">
     import { Transition } from 'shared/components'
-    import { CrashReportingView, LanguageAndAppearanceView, LegalView } from './views'
+    import { LanguageAndAppearanceView, LegalView, WelcomeView } from './views'
     import { _ } from '@core/i18n'
     import { AppSetupRoute, appSetupRoute } from '@core/router'
 </script>
 
-{#if $appSetupRoute === AppSetupRoute.Legal}
+{#if $appSetupRoute === AppSetupRoute.Welcome}
+    <Transition>
+        <WelcomeView />
+    </Transition>
+{:else if $appSetupRoute === AppSetupRoute.Legal}
     <Transition>
         <LegalView />
-    </Transition>
-{:else if $appSetupRoute === AppSetupRoute.CrashReporting}
-    <Transition>
-        <CrashReportingView />
     </Transition>
 {:else if $appSetupRoute === AppSetupRoute.LanguageAndAppearance}
     <Transition>
