--- packages/shared/routes/onboarding/views/app-setup/views/CrashReportingView.svelte
@@ -1,41 +1,0 @@
-<script lang="typescript">
-    import { Animation, Button, Checkbox, OnboardingLayout, Text } from 'shared/components'
-    import { localize } from '@core/i18n'
-    import { appSettings, isAwareOfCrashReporting } from '@core/app'
-    import { appSetupRouter } from '@core/router'
-
-    const busy = false
-    let sendCrashReports = $appSettings?.sendCrashReports ?? true
-
-    const handleBackClick = (): void => {
-        $appSetupRouter.previous()
-    }
-
-    function handleContinueClick(): void {
-        appSettings.set({ ...$appSettings, sendCrashReports })
-
-        if (!$isAwareOfCrashReporting) {
-            isAwareOfCrashReporting.set(true)
-        }
-
-        $appSetupRouter.next()
-    }
-</script>
-
-<OnboardingLayout onBackClick={handleBackClick} {busy}>
-    <div slot="title">
-        <Text type="h2" classes="mb-5">{localize('views.crashReporting.title')}</Text>
-    </div>
-    <div slot="leftpane__content">
-        <Text type="p" secondary classes="mb-8">{localize('views.crashReporting.body')}</Text>
-    </div>
-    <div slot="leftpane__action">
-        <Checkbox label={localize('views.crashReporting.checkbox')} bind:checked={sendCrashReports} classes="mb-8" />
-        <Button classes="w-full" onClick={handleContinueClick}>
-            {localize('actions.continue')}
-        </Button>
-    </div>
-    <div slot="rightpane" class="w-full h-full flex justify-center bg-pastel-orange dark:bg-gray-900">
-        <Animation animation="secure-desktop" />
-    </div>
-</OnboardingLayout>
