+++ packages/shared/routes/onboarding/views/app-setup/views/index.js
@@ -1,3 +1,3 @@
-export { default as CrashReportingView } from './CrashReportingView.svelte'
 export { default as LanguageAndAppearanceView } from './LanguageAndAppearanceView.svelte'
 export { default as LegalView } from './LegalView.svelte'
+export { default as WelcomeView } from './WelcomeView.svelte'
